+++ tests/integration/fixtures/mini_sentry.py
@@ -145,6 +145,7 @@ class Sentry(SentryLike):
                         "$object": ["@password"],
                     },
                 },
+                "transactionMetrics": {"version": 3},
             },
         }